==> code/10_descriptive_analysis.Rmd <==
---
title: "Cardiovascular SSP - Descriptive Analyses"
author: "Sam Jaros"
date: "`r Sys.Date()`"
output: html_document
---

```{r setup}
library(table1)
library(tidyverse)

# Cleaned & merged data import
full_analysis = readRDS(here::here("data/full_analysis.rds"))
```

# Table 1 - Summary of Analysis Variables

```{r table_one}
my_rend = function(x, name, ...) {
  if(is.numeric(x)){
    r = with(stats.default(x), 
             c("", 
               "Mean (SD)"=sprintf("%0.1f (%0.1f)", MEAN, SD),
               "Median (IQR)"=sprintf("%0.1f (%0.1f-%0.1f)", MEDIAN, Q1, Q3)))
  } else {
    r= c("", 
         sapply(stats.default(x), 
                function(y) with(y, sprintf("%d (%0.0f%%)", FREQ, PCT))))
  }
  if(any(is.na(x))){
    r = c(r, with(stats.default(is.na(x))$Yes,
                  c(Missing=sprintf("%d (%0.0f%%)", FREQ, PCT))))
  }
  return(r)
}

# Biological & Demographic Variables
table1(~ hbp1 + sys_bp + dia_bp + 
         age + gender + race + educ +
         diabetes + stroke + mi + htn_med +
         smoke + bmi + hdl + ldl + egfr_ckdepi + crp +
         is.na(geo_tract11) + is.na(geo_zcta5) + is.na(geo_county5) | study,
       data = full_analysis,
       render = my_rend)
```

# Relevant counts

```{r}
# Number in unadjusted tract BP
nrow(drop_na(full_analysis, hbp1, geo_tract11))
# Number in unajdusted tract sys
nrow(drop_na(full_analysis, sys_bp, geo_tract11))
# Number in adjusted tract BP
nrow(drop_na(full_analysis, hbp1, geo_tract11, age, gender, race, educ, 
             diabetes, smoke, mi, bmi, hdl, ldl, egfr_ckdepi, crp))
# Number in adjusted tract sys
nrow(drop_na(full_analysis, sys_bp, geo_tract11, age, gender, race, educ, 
             diabetes, smoke, mi, bmi, hdl, ldl, egfr_ckdepi, crp))

# Number of unique tracts
length(unique(full_analysis$geo_tract11))
# Number of unique ZIP codes
length(unique(full_analysis$geo_zcta5))
# Number of unique counties
length(unique(full_analysis$geo_county5))
```

## ICE Quintile Summary

ICE values defining the quintiles for each geographic level and domain.

```{r}
full_analysis %>%
  select(id, starts_with("ICE")) %>%
  rename_with(~paste(.x, "_value", sep=""), c(everything(),-ends_with("_quint"),-id)) %>%
  pivot_longer(cols=starts_with("ICE"),
               names_sep = "_",
               names_to = c("ICE_var","geo",".value")) %>%
  group_by(ICE_var, geo, quint) %>%
  summarize(ICE_min = round(min(value),2),
            ICE_max = round(max(value),2)) %>%
  filter(geo=="tract")
```
