==> code/12_spatial_modeling_by_race.Rmd <==
---
title: "Cardiovascular SSP - Spatial Modeling By Race"
author: "Sam Jaros"
date: "`r Sys.Date()`"
output: html_document
---

```{r setup}
library(doSNOW)
library(lme4)
library(lmerTest)
library(tidyverse)

# --- Cleaned & merged data import ---------------------------------------------
# Scaling continuous variables due to large eigenvalues
full_analysis = as.data.frame(readRDS(here::here("data/full_analysis.rds")))
full_analysis$age_s = as.vector(scale(full_analysis$age))
full_analysis$bmi_s = as.vector(scale(full_analysis$bmi))
full_analysis$hdl_s = as.vector(scale(full_analysis$hdl))
full_analysis$ldl_s = as.vector(scale(full_analysis$ldl))
full_analysis$egfr_ckdepi_s = as.vector(scale(full_analysis$egfr_ckdepi))
full_analysis$crp_s = as.vector(scale(full_analysis$crp))

# --- Parameters for modeling --------------------------------------------------
outcomes = c("hbp1", "sys_bp")
ICEs = c("ICEincome", "ICEraceeth", "ICEhome", "ICEincwnh", "ICEedu")
ICE_mods = c("", "_quint")
geo_levels = c("county", "zcta", "tract")
model_params = expand.grid("outcome"=outcomes, "ICE"=ICEs, "ICE_mod"=ICE_mods, "geo"=geo_levels,
                           stringsAsFactors=F)

# --- Parameters for saving ----------------------------------------------------
do_save_csv = TRUE
csv_file_name = "results/race_20240222_model.csv"
```

```{r functions}
# Function to run model and return data.frame with outputs
do_mem_race = function(dataset, outcome, control_vars, ICE, ICE_mod, geo_level){
  # Create model formula
  formula = paste0(outcome,
                   "~",control_vars,
                   "race*",ICE,"_",geo_level,ICE_mod,
                   "+(1|geo_",geo_level,ifelse(geo_level=="tract",11,5),")")

  # Run model based on outcome type (numeric = gaussian, factor = binomial)
  if(is.numeric(pull(full_analysis, all_of(outcome)))){
    mod = lmer(formula, data=dataset)
    ci = confint(mod, method="profile")
  } else {
    # Allow for more iterations with the binomial regression
    mod = glmer(formula, data=dataset, family=binomial, 
                control=glmerControl(optimizer="bobyqa", optCtrl=list(maxfun=2e5)))
    # Use Wald confidence intervals to overcome issues when confint would find lower deviances
    # Issues with Wald overcome by large sample size & p~=25%
    ci = confint(mod, method="Wald")
  }

  # Get needed parameters from model
  # Coefficient
  coef = summary(mod)$coefficients
  ICE_coef_names = row.names(coef)[startsWith(rownames(coef), "ICE")]
  ICE_p = unname(coef[startsWith(rownames(coef), "ICE"), startsWith(colnames(coef), "Pr")])
  ICE_est = unname(coef[startsWith(rownames(coef), "ICE"), "Estimate"])
  race_p = unname(coef["raceBlack", startsWith(colnames(coef), "Pr")])
  race_est = unname(coef["raceBlack", "Estimate"])
  int_p = unname(coef[startsWith(rownames(coef), "raceBlack:ICE"), startsWith(colnames(coef), "Pr")])
  int_est = unname(coef[startsWith(rownames(coef), "raceBlack:ICE"), "Estimate"])
  # Confidence interval
  ICE_ci = matrix(ci[startsWith(row.names(ci), "ICE")], ncol=2, byrow=F)
  race_ci = matrix(ci["raceBlack",], ncol=2, byrow=F)
  int_ci = matrix(ci[startsWith(row.names(ci), "raceBlack:ICE")], ncol=2, byrow=F)

  # Assemble and return data
  data.frame("geo"=geo_level, "ICE"=paste0(ICE,ICE_mod), "outcome"=outcome,
             "ICE_coef_name"=ICE_coef_names, "ICE_est"=ICE_est, 
             "ICE_l_ci"=ICE_ci[,1], "ICE_u_ci"=ICE_ci[,2], "ICE_p"=ICE_p,
             "race_est"=race_est, "race_l_ci"=race_ci[,1], 
             "race_u_ci"=race_ci[,2], "race_p"=race_p,
             "int_est"=int_est, "int_l_ci"=int_ci[,1], 
             "int_u_ci"=int_ci[,2], "int_p"=int_p)
}
```

# 1. Basic Model

```{r basic}
# For basic model, just controlling for study
# NOTE, must end with +
basic_control_vars = "study+"

# Setup parallelization
pack_needed = c("dplyr", "lme4", "lmerTest")
cluster = makeCluster(6) # Using 6 of 8 available cores
pb = txtProgressBar(max=nrow(model_params), style=3)
opts = list(progress = function(n) setTxtProgressBar(pb, n))
registerDoSNOW(cluster)

# Run models in parallel
basic_mods = foreach(i=1:nrow(model_params), .combine="rbind", .packages=pack_needed,
                     .options.snow=opts, .errorhandling="stop") %dopar% 
  {
    do_mem_race(full_analysis, model_params[i,1], basic_control_vars, 
                model_params[i,2], model_params[i,3], model_params[i,4])
  }

# Stop parallel clusters
close(pb)
stopCluster(cluster)

basic_mods_wMissing = model_params %>%
  mutate(ICE = paste0(ICE, ICE_mod)) %>%
  select(-ICE_mod) %>%
  full_join(basic_mods, by=c("outcome", "ICE", "geo")) %>%
  mutate(model="basic")

if(do_save_csv){
  write.csv(basic_mods_wMissing, 
            file=here::here("data", str_replace(csv_file_name, "model", "basic")), 
            row.names=F)
}
```

# 2. Demographics Model

```{r demog}
# For demographics model, control for study & demographics
# NOTE, must end with +
demog_control_vars = "study+age_s+gender+educ+"

# Setup parallelization
pack_needed = c("dplyr", "lme4", "lmerTest")
cluster = makeCluster(6) # Using 6 of 8 available cores
pb = txtProgressBar(max=nrow(model_params), style=3)
opts = list(progress = function(n) setTxtProgressBar(pb, n))
registerDoSNOW(cluster)

# Run models in parallel
demo_mods = foreach(i=1:nrow(model_params), .combine="rbind", .packages=pack_needed,
                     .options.snow=opts, .errorhandling="stop") %dopar% 
  {
    do_mem_race(full_analysis, model_params[i,1], demog_control_vars, 
           model_params[i,2], model_params[i,3], model_params[i,4])
  }

# Stop parallel clusters
close(pb)
stopCluster(cluster)

demo_mods_wMissing = model_params %>%
  mutate(ICE = paste0(ICE, ICE_mod)) %>%
  select(-ICE_mod) %>%
  full_join(demo_mods, by=c("outcome", "ICE", "geo")) %>%
  mutate(model="demog")

if(do_save_csv){
  write.csv(demo_mods_wMissing, 
            file=here::here("data", str_replace(csv_file_name, "model", "demo")), 
            row.names=F)
}
```

# 3. Medical History

```{r med_hist}
# For medical history model, control for study & medical history
# NOTE, must end with +
medHist_control_vars = "study+diabetes+stroke+mi+"

# Setup parallelization
pack_needed = c("dplyr", "lme4", "lmerTest")
cluster = makeCluster(6) # Using 6 of 8 available cores
pb = txtProgressBar(max=nrow(model_params), style=3)
opts = list(progress = function(n) setTxtProgressBar(pb, n))
registerDoSNOW(cluster)

# Run models in parallel
medHist_mods = foreach(i=1:nrow(model_params), .combine="rbind", .packages=pack_needed,
                     .options.snow=opts, .errorhandling="stop") %dopar% 
  {
    do_mem_race(full_analysis, model_params[i,1], medHist_control_vars, 
           model_params[i,2], model_params[i,3], model_params[i,4])
  }

# Stop parallel clusters
close(pb)
stopCluster(cluster)

medHist_mods_wMissing = model_params %>%
  mutate(ICE = paste0(ICE, ICE_mod)) %>%
  select(-ICE_mod) %>%
  full_join(medHist_mods, by=c("outcome", "ICE", "geo")) %>%
  mutate(model="medHist")

if(do_save_csv){
  write.csv(medHist_mods_wMissing, 
            file=here::here("data", str_replace(csv_file_name, "model", "medHist")), 
            row.names=F)
}
```

# 4. Comorbididties Model

```{r comorb}
# For demographics model, control for study & demographics
# NOTE, must end with +
comorb_control_vars = "study+smoke+bmi_s+hdl_s+ldl_s+egfr_ckdepi_s+crp_s+"

# Setup parallelization
pack_needed = c("dplyr", "lme4", "lmerTest")
cluster = makeCluster(6) # Using 6 of 8 available cores
pb = txtProgressBar(max=nrow(model_params), style=3)
opts = list(progress = function(n) setTxtProgressBar(pb, n))
registerDoSNOW(cluster)

# Run models in parallel
comorb_mods = foreach(i=1:nrow(model_params), .combine="rbind", .packages=pack_needed,
                     .options.snow=opts, .errorhandling="stop") %dopar% 
  {
    do_mem_race(full_analysis, model_params[i,1], comorb_control_vars, 
           model_params[i,2], model_params[i,3], model_params[i,4])
  }

# Stop parallel clusters
close(pb)
stopCluster(cluster)

comorb_mods_wMissing = model_params %>%
  mutate(ICE = paste0(ICE, ICE_mod)) %>%
  select(-ICE_mod) %>%
  full_join(comorb_mods, by=c("outcome", "ICE", "geo")) %>%
  mutate(model="comorb")

if(do_save_csv){
  write.csv(comorb_mods_wMissing, 
            file=here::here("data", str_replace(csv_file_name, "model", "comorb")), 
            row.names=F)
}
```

# 5. Full Model

```{r full}
# For demographics model, control for study & demographics
# NOTE, must end with +
full_control_vars = "study+age_s+gender+educ+diabetes+smoke+mi+bmi_s+hdl_s+ldl_s+egfr_ckdepi_s+crp_s+"

# Setup parallelization
pack_needed = c("dplyr", "lme4", "lmerTest")
cluster = makeCluster(6) # Using 6 of 8 available cores
pb = txtProgressBar(max=nrow(model_params), style=3)
opts = list(progress = function(n) setTxtProgressBar(pb, n))
registerDoSNOW(cluster)

# Run models in parallel
full_mods = foreach(i=1:nrow(model_params), .combine="rbind", .packages=pack_needed,
                    .options.snow=opts, .errorhandling="stop") %dopar% 
  {
    do_mem_race(full_analysis, model_params[i,1], full_control_vars, 
           model_params[i,2], model_params[i,3], model_params[i,4])
  }

# Stop parallel clusters
close(pb)
stopCluster(cluster)

full_mods_wMissing = model_params %>%
  mutate(ICE = paste0(ICE, ICE_mod)) %>%
  select(-ICE_mod) %>%
  full_join(full_mods, by=c("outcome", "ICE", "geo")) %>%
  mutate(model="full")

if(do_save_csv){
  write.csv(full_mods_wMissing, 
            file=here::here("data", str_replace(csv_file_name, "model", "full")), 
            row.names=F)
}
```

# Combine results

```{r combine}
rbind(basic_mods_wMissing, full_mods_wMissing) %>%
  mutate(across(contains(c("est","l_ci","u_ci")), 
                ~round(if_else(grepl("hbp", outcome), exp(.x) , .x), 4)),
         across(contains("p"),
                ~round(.x, 4))) %>%
  write.csv(here::here("data", str_replace(csv_file_name, "model", "all")), row.names=F)
```

# Get results in table format

```{r make_table_format}
ICE_factor_order = c("ICEedu_quint","ICEhome_quint","ICEincome_quint","ICEraceeth_quint","ICEincwnh_quint")

read_csv(here::here("data","results","race_20240222_all.csv"),
                      show_col_types = F) %>%
  filter(outcome=="hbp1", endsWith(ICE, "_quint"), geo=="tract") %>%
  mutate(quint = str_sub(ICE_coef_name,-1),
         ICE = factor(ICE, levels=ICE_factor_order),
         ICE_col = paste0(round(ICE_est,2)," (",round(ICE_l_ci,2),"-",round(ICE_u_ci,2),")"),
         race_col = paste0(round(race_est,2)," (",round(race_l_ci,2),"-",round(race_u_ci,2),")"),
         int_col = paste0(round(int_est,2)," (",round(int_l_ci,2),"-",round(int_u_ci,2),")")) %>%
  arrange(model, ICE, quint) %>%
  select(model, ICE, quint, ICE_col, race_col, int_col) %>%
  write_csv(file=here::here("data","s_table4.csv"))
```
